requirements.txt
peewee==3.16.3
aiohttp==3.8.5
progressbar2>=4.2.0